requirements.txt
#Web auto-scrape and anit-bot
playwright
playwright-stealth

#data ingestion and processing(images)
pillow
numpy
requests
pdf2image
google-genai

#
thefuzz
boto3